fix: nested fields on find
Fixes #33
