+++ src/commands/mod.rs
@@ -42,7 +42,6 @@ pub use self::list_collections::ListCollections;
 pub use self::list_databases::ListDatabases;
 pub use self::list_indexes::ListIndexes;
 pub use self::ping::Ping;
-pub use self::update::collapse_fields;
 pub use self::update::InvalidUpdateError;
 pub use self::update::Update;
 pub use self::update::UpdateDoc;
