+++ src/commands/update.rs
@@ -1,5 +1,6 @@
 #![allow(dead_code)]
 use crate::handler::{CommandExecutionError, Request};
+use crate::utils::expand_fields;
 use crate::{commands::Handler, pg::SqlParam};
 use bson::{doc, Bson, Document};
 
@@ -97,86 +98,6 @@ impl Handler for Update {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct KeyConflictError {
-    pub source: String,
-    pub target: String,
-}
-
-fn path_to_doc(path: &str, value: &Bson) -> Document {
-    let parts = path.split('.');
-
-    let mut doc = doc! {};
-    let mut first = true;
-    for key in parts.rev() {
-        if first {
-            doc.insert(key, value.clone());
-            first = false;
-        } else {
-            doc = doc! {
-                key: doc
-            };
-        }
-    }
-
-    doc
-}
-
-pub fn expand_fields(doc: &Document) -> Result<Document, KeyConflictError> {
-    let mut expanded = doc![];
-    let mut keys: Vec<&str> = vec![];
-    for (key, value) in doc.iter() {
-        if key.contains(".") {
-            let ikey = key.split(".").next().unwrap();
-            if expanded.contains_key(ikey) {
-                let target = keys
-                    .iter()
-                    .find(|k| {
-                        k.to_string() == ikey.to_string() || k.starts_with(&format!("{}.", ikey))
-                    })
-                    .unwrap();
-                return Err(KeyConflictError {
-                    source: key.to_string(),
-                    target: target.to_string(),
-                });
-            }
-            expanded.insert(ikey, path_to_doc(key, value).get(ikey).unwrap());
-        } else {
-            expanded.insert(key, value);
-        }
-        keys.push(&key);
-    }
-    Ok(expanded)
-}
-
-pub fn collapse_fields(doc: &Document) -> Document {
-    let mut collapsed = doc![];
-    for (key, value) in doc.iter() {
-        if value.as_document().is_none() {
-            collapsed.insert(key, value);
-            continue;
-        }
-
-        let res = collapse_fields(value.as_document().unwrap());
-        for (k, v) in res {
-            collapsed.insert(format!("{}.{}", key, k), v);
-        }
-    }
-    collapsed
-}
-
-fn get_path(doc: &Document, path: String) -> Option<&Bson> {
-    let parts: Vec<&str> = path.split(".").collect();
-    let mut current = doc;
-    for part in parts {
-        match current.get_document(part) {
-            Ok(doc) => current = doc,
-            Err(_) => return current.get(part),
-        }
-    }
-    None
-}
-
 fn parse_update(doc: &Document) -> Result<UpdateOper, InvalidUpdateError> {
     let mut res: Vec<UpdateDoc> = vec![];
     if !doc.keys().any(|k| k.starts_with("$")) {
@@ -253,49 +174,6 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_path_to_doc() {
-        let doc = path_to_doc("a.b.c", &Bson::Int32(1));
-        assert_eq!(
-            doc,
-            doc! {
-                "a": {
-                    "b": {
-                        "c": 1
-                    }
-                }
-            }
-        );
-    }
-
-    #[test]
-    fn test_get_path() {
-        assert_eq!(
-            get_path(&doc! {"x": {"y": {"z": 1}}}, "x.y.z".to_string()).unwrap(),
-            &Bson::Int32(1)
-        );
-        assert_eq!(get_path(&doc! {}, "a.b.c".to_string()), None);
-    }
-
-    #[test]
-    fn test_expand_fields() {
-        let expanded =
-            expand_fields(&doc! { "z": 1, "a.b": 1, "b.c.d": 2, "x.y.z": "Felipe" }).unwrap();
-        assert_eq!(
-            expanded,
-            doc! { "z": 1, "a": { "b": 1 }, "b": { "c": { "d": 2 } }, "x": { "y" : { "z": "Felipe" } } }
-        );
-    }
-
-    #[test]
-    fn test_expand_fields_with_conflict() {
-        let expanded = expand_fields(&doc! { "a.b": 1, "a.b.c": 2 });
-        assert!(expanded.is_err());
-        let err = expanded.unwrap_err();
-        assert_eq!(err.source, "a.b.c");
-        assert_eq!(err.target, "a.b");
-    }
-
-    #[test]
     fn test_parse_update() {
         let set_doc = doc! { "$set": { "a": 1 } };
         let repl_doc = doc! { "b": 2, "c": 8, "d": 9 };
@@ -318,12 +196,5 @@ mod tests {
             parse_update(&mixed_doc).unwrap_err(),
             InvalidUpdateError::new("Unknown modifier: b".to_string())
         );
-    }
-
-    #[test]
-    fn test_collapse_fields() {
-        let nested = doc! { "a": 1, "b": { "c": 2, "d": 3, "e": { "f": 1 } } };
-        let collapsed = collapse_fields(&nested);
-        assert_eq!(collapsed, doc! { "a": 1, "b.c": 2, "b.d": 3, "b.e.f": 1 });
     }
 }
